% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjAF.R
\name{calc_effective_N}
\alias{calc_effective_N}
\title{calc_effective_N}
\usage{
calc_effective_N(N_reference, N_observed, pi.target, pi.observed)
}
\arguments{
\item{N_reference}{numeric vector of the sample sizes of each K reference groups.}

\item{N_observed}{numeric value of the sample size of the observed group.}

\item{pi.target}{numeric vector of the mixture proportions for K reference groups in the target sample or subject. The order must match the order of the reference columns with the last entry matching the missing reference group.}

\item{pi.observed}{numeric vector of the mixture proportions for K reference groups for the observed group. The order must match the order of the reference columns with the last entry matching the missing reference group.}
}
\value{
N_effective: effective sample size for the group that is left out when estimating the adjusted allele frequencies in each adjAF function iteration.
}
\description{
Helper function to calculate effective sample size for the group that is left out when estimating the adjusted allele frequencies in each adjAF function iteration.
}
