% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summix.R
\name{calc_scaledObj}
\alias{calc_scaledObj}
\title{calc_scaledObj}
\usage{
calc_scaledObj(data, reference, observed, pi.start)
}
\arguments{
\item{data}{a dataframe of the observed and reference allele frequencies for N genetic variants. See data formatting document at \href{https://github.com/hendriau/Summix}{https://github.com/hendriau/Summix} for more information. Uses the same input data as summix.}

\item{reference}{a character vector of the column names for the reference groups.}

\item{observed}{a string that is the column name for the observed group.}

\item{pi.start}{Length K numeric vector of the starting guess for the reference group proportions. If not specified, this defaults to 1/K where K is the number of reference groups.}
}
\value{
numeric value that is the scaled objective per 1000 SNPs
}
\description{
Helper function to calculate new scaled loss function using weighted AF bin objectives
}
