% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summix_local.R
\name{doInternalSimulation}
\alias{doInternalSimulation}
\title{doInternalSimulation}
\usage{
doInternalSimulation(windows, data, reference, observed, nRefs, nSim = 1000)
}
\arguments{
\item{windows}{is a dataframe with the Start_Pos and End_Pos}

\item{data}{is the original chromosome data}

\item{reference}{is a list with the names of the columns with references}

\item{observed}{a character value that is the column name for the observed group}

\item{nRefs}{is a vector the same lengths as reference with the number of individuals in each reference population}

\item{nSim}{is the number of internal simulations for the standard error calculations}
}
\description{
Helper function to get the within block se using re-simulation
}
