% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summix_local.R
\name{getNextStartPoint}
\alias{getNextStartPoint}
\title{getNextStartPoint}
\usage{
getNextStartPoint(data, start, end, overlap)
}
\arguments{
\item{data}{the input dataframe subset to the chromosome}

\item{start}{the current index of start point}

\item{end}{the current index of end point}

\item{overlap}{the desired amount of window overlap (in bp or variants)}
}
\value{
returns index of new start point
}
\description{
Helper function: algorithm to get next start point; will pick the point that provides approx. the specified amount of overlap, but not more; if there are only two variants in the previous block, will jump new start point to the previous end point
}
