% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summix_local.R
\name{saveBlock}
\alias{saveBlock}
\title{saveBlock}
\usage{
saveBlock(data, start, end, props, results)
}
\arguments{
\item{data}{the input dataframe subsetting to just the chromosome}

\item{start}{index of start of block}

\item{end}{index of the end of block}

\item{props}{ancestry proportions for the block returned from summix}

\item{results}{current results dataframe}
}
\description{
Helper function to save one block to results
}
