% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summix_local.R
\name{sizeGetNext}
\alias{sizeGetNext}
\title{sizeGetNext}
\usage{
sizeGetNext(positions, start, minSize)
}
\arguments{
\item{positions}{list of positions of variants}

\item{start}{index of the current start position}

\item{minSize}{integer defining the minimum size in bp of the window}
}
\value{
the new end point index
}
\description{
Helper function to get starting end point that is a minimum distance (in bases) from start point; uses indices NOT position numbers
}
