% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summix.R
\name{summix_network}
\alias{summix_network}
\title{summix_network}
\usage{
summix_network(
  data = data,
  sum_res = sum_res,
  reference = reference,
  N_reference = N_reference,
  reference_colors = reference_colors
)
}
\arguments{
\item{data}{A dataframe of the observed and reference allele frequencies for N genetic variants. See data formatting document at \href{https://github.com/hendriau/Summix}{https://github.com/hendriau/Summix} for more information.}

\item{sum_res}{The resulting data frame from the summix function}

\item{reference}{A character vector of the column names for the reference groups.}

\item{N_reference}{numeric vector of the sample sizes for each of the K reference groups.}

\item{reference_colors}{A character vector of length K that specifies the color each reference group node in the network plot. If not specified, this defaults to K random colors.}
}
\value{
network diagram with nodes as estimated substructure proportions and edges as degree of similarity between the given node pair
}
\description{
Helper function to plot the network diagram of estimated substructure proportions and similarity between reference groups
}
