% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summix_local.R
\name{testDiff}
\alias{testDiff}
\title{testDiff}
\usage{
testDiff(last, current, threshold = 0.01)
}
\arguments{
\item{last}{substructure proportions of block returned from summix}

\item{current}{substructure proportions of block returned from summix}

\item{threshold}{if applicable the threshold for determining change point}
}
\value{
true if passes threshold, false if not
}
\description{
Helper function to determine whether reference group has changed for fast/catchup window algorithm
}
