% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summix_local.R
\name{variantGetNext}
\alias{variantGetNext}
\title{variantGetNext}
\usage{
variantGetNext(positions, start, minVariants)
}
\arguments{
\item{positions}{list of positions of variants}

\item{start}{index of the current start position}

\item{minVariants}{integer defining the minimum size in number of variants of the window}
}
\value{
the new end point index
}
\description{
Helper function to get starting end point that is a minimum distance (in variants) from start point; uses indices NOT position numbers
}
